third_party/WebKit/LayoutTests/intersection-observer/edge-inclusive-intersection-expected.txt

Verifies that IntersectionObserver detects edge-inclusive intersections.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS entries.length is 0
PASS entries.length is 0
PASS entries.length is 1
PASS entries[0].boundingClientRect.left is 8
PASS entries[0].boundingClientRect.right is 108
PASS entries[0].boundingClientRect.top is 208
PASS entries[0].boundingClientRect.bottom is 308
PASS entries[0].intersectionRect.left is 8
PASS entries[0].intersectionRect.right is 108
PASS entries[0].intersectionRect.top is 208
PASS entries[0].intersectionRect.bottom is 208
PASS entries[0].rootBounds.left is 8
PASS entries[0].rootBounds.right is 208
PASS entries[0].rootBounds.top is 8
PASS entries[0].rootBounds.bottom is 208
PASS entries[0].target is [object HTMLDivElement]
PASS entries.length is 2
PASS entries.length is 3
PASS entries[2].boundingClientRect.left is 8
PASS entries[2].boundingClientRect.right is 308
PASS entries[2].boundingClientRect.top is 193
PASS entries[2].boundingClientRect.bottom is 193
PASS entries[2].intersectionRect.left is 8
PASS entries[2].intersectionRect.right is 208
PASS entries[2].intersectionRect.top is 193
PASS entries[2].intersectionRect.bottom is 193
PASS entries[2].rootBounds.left is 8
PASS entries[2].rootBounds.right is 208
PASS entries[2].rootBounds.top is 8
PASS entries[2].rootBounds.bottom is 208
PASS entries[2].target is [object HTMLDivElement]
PASS successfullyParsed is true

TEST COMPLETE
